feat: refactor monolithic xaml-converter into focused modules
- Split 1,187-line XamlToMarkdownConverter into 4 focused modules
- Created XamlElementProcessor for element processing logic
- Created XamlFormattingService for text formatting operations
- Created XamlUtilities for validation and utility functions
- Reduced main converter from 1,187 to 208 lines (83% reduction)
- Maintained full API compatibility and functionality
